MIT App Inventor 2 Bluetooth Robot/LED Control App
My attempt to copy some of the functionality of Joystick Bluetooth
Commander. My version is not nearly as nice as the original.

https://play.google.com/store/apps/details?id=org.projectproto.btjoystick
I added a screen which should allow some control over an 8x8 LED matrix
similar to this once from Adafruit.
https://www.adafruit.com/products/1487
There are three YouTube videos describing aspects of App Inventor 2 and
aspects of my app.
The Propeller Activity Board had an experimental version of the Eddie
firmware loaded.
I'll add links to the YouTube videos and a link to the Propeller object
once I upload these files. These added links should be in the readme
file once they're ready.